feat(api/v1): LaxInt32 — accept CLI-quoted integers in JSON
The upstream `linstor` Python client marshals integer flags as
JSON-quoted strings (e.g. `"place_count":"2"`) even though the OpenAPI
spec types them as int32. Go's strict decoder rejects this with
`cannot unmarshal string into Go struct field … of type int32`, so
every `linstor rg create --place-count N` died at the API layer.

LaxInt32 accepts both `2` and `"2"` on unmarshal and emits the bare
integer form on marshal — so a request round-tripped through blockstor
matches the spec rather than propagating the CLI's quirk.

AutoSelectFilter.PlaceCount and AdditionalPlaceCount switch to
LaxInt32 with explicit casts at the wire↔CRD boundaries (CRD keeps
int32 — only the wire decoder needs the leniency).

+++ pkg/rest/advise.go
@@ -120,7 +120,7 @@ func (s *Server) adviseOne(ctx context.Context, rd *apiv1.ResourceDefinition) (a
 
 	entry := adviceEntry{
 		Name:        rd.Name,
-		PlaceCount:  filter.PlaceCount,
+		PlaceCount:  int32(filter.PlaceCount),
 		Suggestions: picks,
 	}
